feat: global guard, public, apply task

+++ src/task/task.controller.ts
@@ -6,33 +6,62 @@ import {
   Patch,
   Param,
   Delete,
+  ForbiddenException,
 } from '@nestjs/common';
 import { TaskService } from './task.service';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
+import { ActiveUser } from 'src/auth/activeUser.decorator';
+import { Task } from './entities/task.entity';
 
 @Controller('task')
 export class TaskController {
   constructor(private readonly taskService: TaskService) {}
 
   @Post()
-  async create(@Body() createTaskDto: CreateTaskDto) {
-    const createdTask = await this.taskService.create(createTaskDto);
+  async create(
+    @Body() createTaskDto: CreateTaskDto,
+    @ActiveUser('sub') userId: number,
+  ) {
+    const createdTask = await this.taskService.create({
+      ...createTaskDto,
+      createdBy: userId,
+    });
     return createdTask;
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateTaskDto: UpdateTaskDto) {
-    return this.taskService.update(+id, updateTaskDto);
+  async update(
+    @Param('id') id: string,
+    @Body() updateTaskDto: UpdateTaskDto,
+    @ActiveUser('sub') userId: number,
+  ): Promise<Task> {
+    const task = await this.taskService.findOne(+id);
+
+    if (task.createdBy !== userId) {
+      throw new ForbiddenException();
+    }
+    return await this.taskService.update(+id, updateTaskDto);
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.taskService.findOne(+id);
+  async findOne(
+    @Param('id') id: string,
+    @ActiveUser('sub') userId: number,
+  ): Promise<Task> {
+    const task = await this.taskService.findOne(+id);
+    if (task.createdBy !== userId) {
+      throw new ForbiddenException();
+    }
+    return task;
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string, @ActiveUser('sub') userId: number) {
+    const task = await this.taskService.findOne(+id);
+    if (task.createdBy !== userId) {
+      throw new ForbiddenException();
+    }
     return this.taskService.remove(+id);
   }
 }
